Fix active_shards_percent_as_number type to float

+++ pkg/elasticsearch/client.go
@@ -147,20 +147,20 @@ func (c *clientImpl) GetIndexSettings(ctx context.Context, indexName string) (*I
 }
 
 type IndexHealth struct {
-	ClusterName                 string `json:"cluster_name"`
-	Status                      string `json:"status"`
-	NumberOfNodes               int    `json:"number_of_nodes"`
-	NumberOfDataNodes           int    `json:"number_of_data_nodes"`
-	ActivePrimaryShards         int    `json:"active_primary_shards"`
-	ActiveShards                int    `json:"active_shards"`
-	RelocatingShards            int    `json:"relocating_shards"`
-	InitializingShards          int    `json:"initializing_shards"`
-	UnassignedShards            int    `json:"unassigned_shards"`
-	DelayedUnassignedShards     int    `json:"delayed_unassigned_shards"`
-	NumberOfPendingTasks        int    `json:"number_of_pending_tasks"`
-	NumberOfInFlightFetch       int    `json:"number_of_in_flight_fetch"`
-	TaskMaxWaitingInQueueMillis int    `json:"task_max_waiting_in_queue_millis"`
-	ActiveShardsPercentAsNumber int    `json:"active_shards_percent_as_number"`
+	ClusterName                 string  `json:"cluster_name"`
+	Status                      string  `json:"status"`
+	NumberOfNodes               int     `json:"number_of_nodes"`
+	NumberOfDataNodes           int     `json:"number_of_data_nodes"`
+	ActivePrimaryShards         int     `json:"active_primary_shards"`
+	ActiveShards                int     `json:"active_shards"`
+	RelocatingShards            int     `json:"relocating_shards"`
+	InitializingShards          int     `json:"initializing_shards"`
+	UnassignedShards            int     `json:"unassigned_shards"`
+	DelayedUnassignedShards     int     `json:"delayed_unassigned_shards"`
+	NumberOfPendingTasks        int     `json:"number_of_pending_tasks"`
+	NumberOfInFlightFetch       int     `json:"number_of_in_flight_fetch"`
+	TaskMaxWaitingInQueueMillis int     `json:"task_max_waiting_in_queue_millis"`
+	ActiveShardsPercentAsNumber float64 `json:"active_shards_percent_as_number"`
 }
 
 func (i *IndexHealth) IsHealthy() bool {
